requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
django==4.2.26
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.11
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.4
sqlparse==0.5.4
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
